lowered journal parsing memory usage
