+++ .changes/unreleased/Changed-20250611-214652.yaml
@@ -1,0 +1,3 @@
+kind: Fixed
+body: Lowered memory usage when parsing linux journal files
+time: 2025-06-11T21:46:52.748381596-04:00
