+++ .changes/unreleased/Changed-20250611-214728.yaml
@@ -1,0 +1,3 @@
+kind: Changed
+body: Faster processing when compressing json output
+time: 2025-06-11T21:47:28.806789268-04:00
