+++ Cargo.toml
@@ -28,7 +28,7 @@ base64 = "0.22.1"
 tokio = { version = "1.45.1", features = ["full"] }
 flate2 = "1.1.2"
 glob = "0.3.2"
-reqwest = { version = "0.12.19", features = [
+reqwest = { version = "0.12.20", features = [
     "json",
     "blocking",
     "native-tls-vendored",
