+++ forensics/Cargo.toml
@@ -47,7 +47,7 @@ ruzstd = "0.8.1"
 lz4_flex = "0.11.3"
 xz2 = { version = "0.1.7", default-features = false, features = ["static"] }
 macos-unifiedlogs = "0.2.0"
-plist = "1.7.1"
+plist = "1.7.2"
 aes = "0.8.4"
 cbc = "0.1.2"
 yara-x = { version = "1.0.1", optional = true, default-features = false, features = [
@@ -70,7 +70,7 @@ url = "2.5.4"
 common = { path = "../common" }
 timeline = { path = "../timeline" }
 sunlight = "0.1.1"
-miniz_oxide = { version = "0.8.8", features = ["std", "with-alloc"] }
+miniz_oxide = { version = "0.8.9", features = ["std", "with-alloc"] }
 lumination = "0.1.1"
 
 # JS API
