+++ forensics/src/artifacts/os/linux/journals/error.rs
@@ -5,7 +5,6 @@ pub(crate) enum JournalError {
     SeekError,
     ReadError,
     ObjectHeader,
-    Serialize,
     ReaderError,
     JournalHeader,
     NotJournal,
@@ -20,7 +19,6 @@ impl fmt::Display for JournalError {
             JournalError::ReadError => write!(f, "Failed to read journal data"),
             JournalError::ObjectHeader => write!(f, "Failed to parse object header"),
             JournalError::JournalHeader => write!(f, "Failed to parse journal header"),
-            JournalError::Serialize => write!(f, "Could not serialize data"),
             JournalError::ReaderError => write!(f, "Could not create reader"),
             JournalError::NotJournal => write!(f, "Not a journal file"),
         }
